feat(m19-g5): implement view model layer retrofit — extract pure functions to trajectoryViewModel.ts (#1522)
- Creates `trajectoryViewModel.ts` with MergedStepDatum, computeYDomain, computeDivergenceFill,
  getConfidenceBadgeVisible, mergeTrajectories (+ visibleStepRange param), sliceToStepRange
- Updates TrajectoryView.tsx to import from new module; re-exports all symbols for backward compat
- visibleStepRange seam enables #1524 trackwheel zoom without further TrajectoryView changes
- All 33 unit tests GREEN (trajectoryViewModel.test.ts)
